+++ changes.md
@@ -27,6 +27,14 @@
 - added method `apiRoot.withProjectKey().inStoreKeyWithStoreKeyValue().discountCodes().withId().head()`
 </details>
 
+
+<details>
+<summary>Added Enum(s)</summary>
+
+- added enum `variant` to type `AttributeReferenceTypeId`
+- added enum `variant` to type `CustomFieldReferenceValue`
+</details>
+
 **History changes**
 
 <details>